requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
idna==3.7
python-dotenv==1.0.1
requests==2.32.3
setuptools==69.5.1
sqlparse==0.5.0
urllib3==2.2.1
wheel==0.43.0
